Rename Registry to MatchRegistry
Use a less generic name to make room for an IdentityRegistry.
